Moved functions to more sensible locations
