+++ mcmcfit.py
@@ -14,171 +14,14 @@ import configobj
 import emcee
 import matplotlib.pyplot as plt
 import numpy as np
-from func_timeout import FunctionTimedOut, func_timeout
 
 import mcmc_utils as utils
-from CVModel import (Band, ComplexEclipse, ComplexGPEclipse, GPLCModel,
-                     LCModel, Lightcurve, SimpleEclipse, SimpleGPEclipse)
-from mcmc_utils import Param
 import plot_model_output as summariser
-
-
-def extract_par_and_key(key):
-    '''As stated. For example,
-    extract_par_and_key("wdFlux_long_complex_key_label)
-    >>> ("wdFlux", "long_complex_key_label")
-    '''
-
-    if key.startswith("ln_"):
-        key = key.split("_")
-
-        par = "_".join(key[:3])
-        label = "_".join(key[3:])
-
-    else:
-        par, label = key.split('_')[0], '_'.join(key.split('_')[1:])
-
-    return par, label
-
-def construct_model(input_file):
-    '''Takes an input filename, and parses it into a model tree.
-    Returns that model tree. '''
-
-    input_dict = configobj.ConfigObj(input_file)
-
-    for key, item in input_dict.items():
-        if key in ['ampin_gp', 'ampout_gp', 'tau_gp']:
-            print("OI! Don't use {0}, use ln_{0}!!".format(key))
-            input_dict['ln_'+key] = item
-            input("As punishment, you have to hit enter to continue:\n> ")
-
-    # Do we use the complex model? Do we use the GP?
-    is_complex = bool(int(input_dict['complex']))
-    use_gp = bool(int(input_dict['useGP']))
-
-    # neclipses no longer strictly necessary, but can be used to limit the
-    # maximum number of fitted eclipses
-    try:
-        neclipses = int(input_dict['neclipses'])
-    except KeyError:
-        # Read in all the available eclipses
-        neclipses = 9999
-
-    # # # # # # # # # # # # # # # # #
-    # Get the initial model setup # #
-    # # # # # # # # # # # # # # # # #
-
-    # Start by creating the overall Model. Gather the parameters:
-    if use_gp:
-        core_par_names = GPLCModel.node_par_names
-        core_pars = [Param.fromString(name, input_dict[name])
-                     for name in core_par_names]
-
-        # and make the model object with no children
-        model = GPLCModel('core', core_pars)
-    else:
-        core_par_names = LCModel.node_par_names
-        core_pars = [Param.fromString(name, input_dict[name])
-                     for name in core_par_names]
-
-        # and make the model object with no children
-        model = LCModel('core', core_pars)
-
-    # # # # # # # # # # # # # # # # #
-    # # # Now do the band names # # #
-    # # # # # # # # # # # # # # # # #
-
-    # Collect the bands and their params. Add them total model.
-    band_par_names = Band.node_par_names
-
-    # Use the Eclipse class to find the parameters we're interested in
-    if is_complex:
-        ecl_pars = ComplexEclipse.node_par_names
-    else:
-        ecl_pars = SimpleEclipse.node_par_names
-
-    # I care about the order in which eclipses and bands are defined.
-    # Collect that order here.
-    defined_bands = []
-    defined_eclipses = []
-    with open(input_file, 'r') as input_file_obj:
-        for line in input_file_obj:
-            line = line.strip().split()
-
-            if len(line):
-                key = line[0]
-
-                # Check that the key starts with any of the band pars
-                if np.any([key.startswith(par) for par in band_par_names]):
-                    # Strip off the variable part, and keep only the label
-                    _, key = extract_par_and_key(key)
-                    if key not in defined_bands:
-                        defined_bands.append(key)
-
-                # Check that the key starts with any of the eclipse pars
-                if np.any([key.startswith(par) for par in ecl_pars]):
-                    # Strip off the variable part, and keep only the label
-                    _, key = extract_par_and_key(key)
-                    if key not in defined_eclipses:
-                        defined_eclipses.append(key)
-
-    # Collect the band params into their Band objects.
-    for label in defined_bands:
-        band_pars = []
-
-        # Build the Param objects for this band
-        for par in band_par_names:
-                # Construct the parameter key and retrieve the string
-                key = "{}_{}".format(par, label)
-                string = input_dict[key]
-
-                # Make the Param object, and save it
-                band_pars.append(Param.fromString(par, string))
-
-        # Define the band as a child of the model.
-        Band(label, band_pars, parent=model)
-
-    # # # # # # # # # # # # # # # # #
-    # # Finally, get the eclipses # #
-    # # # # # # # # # # # # # # # # #
-
-    lo = float(input_dict['phi_start'])
-    hi = float(input_dict['phi_end'])
-
-    for label in defined_eclipses[:neclipses]:
-        # Get the list of parameters, and their priors
-        params = []
-        for par in ecl_pars:
-            key = "{}_{}".format(par, label)
-            param = Param.fromString(par, input_dict[key])
-
-            params.append(param)
-
-        # Get the observational data
-        lc_fname = input_dict['file_{}'.format(label)]
-        lc = Lightcurve.from_calib(lc_fname)
-        lc.trim(lo, hi)
-
-        # Get the band object that this eclipse belongs to
-        my_band = input_dict['band_{}'.format(label)]
-        my_band = model.search_Node('Band', my_band)
-
-        if use_gp:
-            if is_complex:
-                ComplexGPEclipse(lc, label, params, parent=my_band)
-            else:
-                SimpleGPEclipse(lc, label, params, parent=my_band)
-        else:
-            if is_complex:
-                ComplexEclipse(lc, label, params, parent=my_band)
-            else:
-                SimpleEclipse(lc, label, params, parent=my_band)
-
-    # Make sure that all the model's Band have eclipses. Otherwise, prune them
-    model.children = [band for band in model.children if len(band.children)]
-
-    return model
-
+from CVModel import (Band, ComplexEclipse, ComplexGPEclipse, GPLCModel,
+                     LCModel, Lightcurve, SimpleEclipse, SimpleGPEclipse,
+                     construct_model, extract_par_and_key)
+from func_timeout import FunctionTimedOut, func_timeout
+from mcmc_utils import Param
 
 if __name__ in '__main__':
